Rename components to -names style
